requirements.txt
praw==7.7.1  
python-dotenv==1.0.0  
apscheduler==3.10.4  
discord-webhook==1.2.1  
huggingface_hub==0.22.2  